Allow russian doll caching in the view
New cacheable mixin to allow components to set a cache key.  This allows
the view to check each component for a cache key and then check the
cache if a key exists.  This will allow different keys through the view
for russian doll caching.

This is a WIP spike with further improvements planned:
- redis-sinatra is a very basic implementation, but worked for POC.  I'd
like to swap this out for sinatra-redis-cache or better yet have a cache
function set via config.  That would also allow all caching to be
disable by setting the cache_func to nil.
- Add a global cache key on config.  Changing this key will allow
invalidation of all caches during a deploy when a global ui/css change
has been made.
- Other components need to be looked at to decide if they should include
Cacheable and then edit their erb to handle caching.